build: use same test program for two tests
